MacGui: remove a useless cast for the deblock value.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@7251 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ macosx/HBFilters.h
@@ -32,7 +32,7 @@ extern NSString * const HBFiltersChangedNotification;
 @property (nonatomic, readwrite, copy) NSString *denoiseTune;
 @property (nonatomic, readwrite, copy) NSString *denoiseCustomString;
 
-@property (nonatomic, readwrite) NSInteger deblock;
+@property (nonatomic, readwrite) int deblock;
 @property (nonatomic, readwrite) BOOL grayscale;
 
 @end
